add requestToJoin method
